Multi-thread feature activated
in main.py, if the _query argument is listed as an SQL query, loop_query_pool has been changed to accept a list argument named 'query' and run as a multithreaded function.
In DB.py, pg_query_pool uses the generator to unpack the list. The unpacked list is a query and each query runs as a connection.